refactor(app): move factory to stripe app
