+++ openmeter/appstripe/adapter/adapter.go
@@ -7,17 +7,23 @@ import (
 	"fmt"
 
 	"github.com/openmeterio/openmeter/openmeter/app"
+	appentity "github.com/openmeterio/openmeter/openmeter/app/entity"
 	"github.com/openmeterio/openmeter/openmeter/appstripe"
+	appstripeentity "github.com/openmeterio/openmeter/openmeter/appstripe/entity"
 	"github.com/openmeterio/openmeter/openmeter/customer"
 	entdb "github.com/openmeterio/openmeter/openmeter/ent/db"
+	"github.com/openmeterio/openmeter/openmeter/secret"
 	"github.com/openmeterio/openmeter/pkg/framework/entutils"
 	"github.com/openmeterio/openmeter/pkg/framework/transaction"
 )
 
 type Config struct {
-	Client          *entdb.Client
-	AppService      app.Service
-	CustomerService customer.Service
+	Client              *entdb.Client
+	AppService          app.Service
+	CustomerService     customer.Service
+	Marketplace         app.MarketplaceService
+	SecretService       secret.Service
+	StripeClientFactory func(apiKey string) StripeClient
 }
 
 func (c Config) Validate() error {
@@ -33,6 +39,14 @@ func (c Config) Validate() error {
 		return errors.New("customer service is required")
 	}
 
+	if c.Marketplace == nil {
+		return errors.New("marketplace is required")
+	}
+
+	if c.SecretService == nil {
+		return errors.New("secret service is required")
+	}
+
 	return nil
 }
 
@@ -45,6 +59,30 @@ func New(config Config) (appstripe.Adapter, error) {
 		db:              config.Client,
 		appService:      config.AppService,
 		customerService: config.CustomerService,
+	}
+
+	stripeClientFactory := config.StripeClientFactory
+	if stripeClientFactory == nil {
+		stripeClientFactory = StripeClientFactory
+	}
+
+	stripeAppFactory, err := NewAppFactory(AppFactoryConfig{
+		AppService:          config.AppService,
+		AppStripeAdapter:    adapter,
+		Client:              config.Client,
+		SecretService:       config.SecretService,
+		StripeClientFactory: stripeClientFactory,
+	})
+	if err != nil {
+		return adapter, fmt.Errorf("failed to create stripe app factory: %w", err)
+	}
+
+	err = config.Marketplace.Register(appentity.RegistryItem{
+		Listing: appstripeentity.StripeMarketplaceListing,
+		Factory: stripeAppFactory,
+	})
+	if err != nil {
+		return adapter, fmt.Errorf("failed to register stripe app: %w", err)
 	}
 
 	return adapter, nil
